Emissions update
1. Refactoring Emissions package
2. Change ParticleEmisiion from interface to abstract class: Pull Up fields, methods, Constructor Body
3. Almost all subclasses had same field, methods with same results and partially same constructors. Those who have not just override the method
