+++ src/main/java/project_16x16/particleSystem/emissions/ArchEmission.java
@@ -1,10 +1,7 @@
 package project_16x16.particleSystem.emissions;
 
 import java.util.Random;
-import java.util.function.Consumer;
-
 import processing.core.PVector;
-import project_16x16.particleSystem.Particle;
 
 /**
  * Arch Emission
@@ -14,18 +11,10 @@ import project_16x16.particleSystem.Particle;
  *
  * @author petturtle
  */
-public class ArchEmission implements ParticleEmission {
+public class ArchEmission extends ParticleEmission {
 
-	private PVector position;
-	private float velocity;
-	private float acceleration;
-	private float spread;
 	private float minAngle;
 	private float maxAngle;
-	
-	private PVector newPosition;
-	private PVector newVelocity;
-	private PVector newAcceleration;
 	
 	/**
      * Create a new ArchEmission.
@@ -38,10 +27,7 @@ public class ArchEmission implements ParticleEmission {
      * @param maxAngle	   maxAngle (radians)
      */
 	public ArchEmission(PVector position, float velocity, float acceleration, float spread, float minAngle, float maxAngle) {
-		this.position = position;
-		this.velocity = velocity;
-		this.acceleration = acceleration;
-		this.spread = spread;
+		super(position, velocity, acceleration, spread);
 		this.minAngle = minAngle;
 		this.maxAngle = maxAngle;
 	}
@@ -52,41 +38,6 @@ public class ArchEmission implements ParticleEmission {
 		newPosition();
 		newVelocity(phi);
 		newAcceleration(phi);
-	}
-	
-	private void newPosition() {
-		PVector p = position.copy();
-		Random ran = new Random();
-		p.x += (ran.nextFloat()*spread*2f)-spread;
-		p.y += (ran.nextFloat()*spread*2f)-spread;
-		newPosition = p;
-	}
-
-	private void newVelocity(float phi) {
-		newVelocity = new PVector();
-		newVelocity.x = (float) (velocity*Math.cos(phi));
-		newVelocity.y = (float) (velocity*Math.sin(phi));
-	}
-
-	private void newAcceleration(float phi) {
-		newAcceleration = new PVector();
-		newAcceleration.x = (float) (acceleration*Math.cos(phi));
-		newAcceleration.y = (float) (acceleration*Math.sin(phi));
-	}
-	
-	@Override
-	public Consumer<Particle> getConsumer() {
-		return p -> {
-			generateNew();
-			p.position = newPosition;
-			p.velocity = newVelocity;
-			p.acceleration = newAcceleration;
-		};
-	}
-
-	@Override
-	public void setPosition(PVector position) {
-		this.position = position;
 	}
 
 	@Override
