+++ src/main/java/project_16x16/particleSystem/emissions/AreaEmission.java
@@ -1,10 +1,6 @@
 package project_16x16.particleSystem.emissions;
 
-import java.util.Random;
-import java.util.function.Consumer;
-
 import processing.core.PVector;
-import project_16x16.particleSystem.Particle;
 
 /**
  * AreaEmission
@@ -13,17 +9,7 @@ import project_16x16.particleSystem.Particle;
  *
  * @author petturtle
  */
-public class AreaEmission implements ParticleEmission {
-
-	private PVector position;
-	private float velocity;
-	private float acceleration;
-	private float spread;
-	
-	private PVector newPosition;
-	private PVector newVelocity;
-	private PVector newAcceleration;
-	
+public class AreaEmission extends ParticleEmission {
 	/**
      * Create a new AreaEmission.
 
@@ -33,10 +19,7 @@ public class AreaEmission implements ParticleEmission {
      * @param spread	   Deviation from spawn position
      */
 	public AreaEmission(PVector position, float velocity, float acceleration, float spread) {
-		this.position = position;
-		this.velocity = velocity;
-		this.acceleration = acceleration;
-		this.spread = spread;
+		super(position, velocity, acceleration, spread);
 	}
 	
 	public void generateNew() {
@@ -46,40 +29,6 @@ public class AreaEmission implements ParticleEmission {
 		newAcceleration(phi);
 	}
 	
-	private void newPosition() {
-		PVector p = position.copy();
-		Random ran = new Random();
-		p.x += (ran.nextFloat()*spread*2f)-spread;
-		p.y += (ran.nextFloat()*spread*2f)-spread;
-		newPosition = p;
-	}
-
-	private void newVelocity(float phi) {
-		newVelocity = new PVector();
-		newVelocity.x = (float) (velocity*Math.cos(phi));
-		newVelocity.y = (float) (velocity*Math.sin(phi));
-	}
-
-	private void newAcceleration(float phi) {
-		newAcceleration = new PVector();
-		newAcceleration.x = (float) (acceleration*Math.cos(phi));
-		newAcceleration.y = (float) (acceleration*Math.sin(phi));
-	}
-	
-	@Override
-	public Consumer<Particle> getConsumer() {
-		return p -> {
-			generateNew();
-			p.position = newPosition;
-			p.velocity = newVelocity;
-			p.acceleration = newAcceleration;
-		};
-	}
-	
-	@Override
-	public void setPosition(PVector position) {
-		this.position = position;
-	}
 
 	@Override
 	public ParticleEmission copy() {
