+++ src/main/java/project_16x16/particleSystem/emissions/RectEmission.java
@@ -1,10 +1,7 @@
 package project_16x16.particleSystem.emissions;
 
 import java.util.Random;
-import java.util.function.Consumer;
-
 import processing.core.PVector;
-import project_16x16.particleSystem.Particle;
 
 /**
  * RectEmission
@@ -13,17 +10,10 @@ import project_16x16.particleSystem.Particle;
  *
  * @author petturtle
  */
-public class RectEmission implements ParticleEmission {
+public class RectEmission extends ParticleEmission {
 
-	private PVector position;
-	private float velocity;
-	private float acceleration;
 	private int width;
 	private int height;
-	
-	private PVector newPosition;
-	private PVector newVelocity;
-	private PVector newAcceleration;
 	
 	/**
      * Create a new AreaEmission.
@@ -35,9 +25,7 @@ public class RectEmission implements ParticleEmission {
      * @param height	   height of rect
      */
 	public RectEmission(PVector position, float velocity, float acceleration, int width, int height) {
-		this.position = position;
-		this.velocity = velocity;
-		this.acceleration = acceleration;
+		super(position, velocity, acceleration, 0);
 		this.width = width;
 		this.height = height;
 	}
@@ -49,39 +37,13 @@ public class RectEmission implements ParticleEmission {
 		newAcceleration(phi);
 	}
 	
-	private void newPosition() {
+	@Override
+	public void newPosition() {
 		PVector p = position.copy();
 		Random ran = new Random();
 		p.x += (ran.nextFloat()*width/2f)-width/2;
 		p.y += (ran.nextFloat()*height/2f)-height/2;
-		newPosition = p;
-	}
-
-	private void newVelocity(float phi) {
-		newVelocity = new PVector();
-		newVelocity.x = (float) (velocity*Math.cos(phi));
-		newVelocity.y = (float) (velocity*Math.sin(phi));
-	}
-
-	private void newAcceleration(float phi) {
-		newAcceleration = new PVector();
-		newAcceleration.x = (float) (acceleration*Math.cos(phi));
-		newAcceleration.y = (float) (acceleration*Math.sin(phi));
-	}
-	
-	@Override
-	public Consumer<Particle> getConsumer() {
-		return p -> {
-			generateNew();
-			p.position = newPosition;
-			p.velocity = newVelocity;
-			p.acceleration = newAcceleration;
-		};
-	}
-	
-	@Override
-	public void setPosition(PVector position) {
-		this.position = position;
+		setPosition(p);
 	}
 	
 	@Override
