+++ src/main/java/project_16x16/particleSystem/emissions/RotationEmission.java
@@ -1,10 +1,6 @@
 package project_16x16.particleSystem.emissions;
 
-import java.util.Random;
-import java.util.function.Consumer;
-
 import processing.core.PVector;
-import project_16x16.particleSystem.Particle;
 
 /**
  * RotationEmission
@@ -15,19 +11,11 @@ import project_16x16.particleSystem.Particle;
  *
  * @author petturtle
  */
-public class RotationEmission implements ParticleEmission {
+public class RotationEmission extends ParticleEmission {
 
-	private PVector position;
-	private float velocity;
-	private float acceleration;
-	private float spread;
 	private float div;
 	private float phi;
 	
-	
-	private PVector newPosition;
-	private PVector newVelocity;
-	private PVector newAcceleration;
 	
 	/**
      * Create a new RotationEmission.
@@ -39,10 +27,7 @@ public class RotationEmission implements ParticleEmission {
      * @param div		   Angle increase for each particle (radians)
      */
 	public RotationEmission(PVector position, float velocity, float acceleration, float spread, float div) {
-		this.position = position;
-		this.velocity = velocity;
-		this.acceleration = acceleration;
-		this.spread = spread;
+		super(position, velocity, acceleration, spread);
 		this.div = div;
 	}
 
@@ -51,41 +36,6 @@ public class RotationEmission implements ParticleEmission {
 		newPosition();
 		newVelocity(phi);
 		newAcceleration(phi);
-	}
-	
-	private void newPosition() {
-		PVector p = position.copy();
-		Random ran = new Random();
-		p.x += (ran.nextFloat()*spread*2f)-spread;
-		p.y += (ran.nextFloat()*spread*2f)-spread;
-		newPosition = p;
-	}
-
-	private void newVelocity(float phi) {
-		newVelocity = new PVector();
-		newVelocity.x = (float) (velocity*Math.cos(phi));
-		newVelocity.y = (float) (velocity*Math.sin(phi));
-	}
-
-	private void newAcceleration(float phi) {
-		newAcceleration = new PVector();
-		newAcceleration.x = (float) (acceleration*Math.cos(phi));
-		newAcceleration.y = (float) (acceleration*Math.sin(phi));
-	}
-
-	@Override
-	public Consumer<Particle> getConsumer() {
-		return p -> {
-			generateNew();
-			p.position = newPosition;
-			p.velocity = newVelocity;
-			p.acceleration = newAcceleration;
-		};
-	}
-	
-	@Override
-	public void setPosition(PVector position) {
-		this.position = position;
 	}
 	
 	@Override
